RegistController 구현 & RegistService 리팩토링

+++ src/main/java/com/samsung/guesting/service/RegistService.java
@@ -1,19 +1,18 @@
 package com.samsung.guesting.service;
 
+import java.time.LocalDateTime;
 import java.util.List;
-import java.util.stream.Collector;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 
-import com.samsung.guesting.dto.RegistReq;
 import com.samsung.guesting.dto.RegistRes;
 import com.samsung.guesting.entity.Regist;
 import com.samsung.guesting.entity.Team;
 import com.samsung.guesting.repository.MemberRepository;
 import com.samsung.guesting.repository.RegistRepository;
 
-import jakarta.servlet.http.HttpSession;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 
@@ -22,37 +21,49 @@ import lombok.RequiredArgsConstructor;
 public class RegistService {
 	private final RegistRepository registRepository;
 	private final MemberRepository memberRepository;
-	private final HttpSession session;
 	
 	//1. 게스팅 신청 : 상대 팀 아이디를 받아서 게스팅을 신청한다.
-	public Regist requestGuesting(RegistReq registReq, Integer curTeamId) {
-		registReq.setSendTeamId(curTeamId); 
+	public Regist requestGuesting(Integer memberId, Integer otherTeamId) {
+		Team curTeam = memberRepository.findByMemberId(memberId)
+				.orElseThrow(() -> new RuntimeException("다시 로그인 해주세요.")).getTeam();
+		if (Objects.isNull(curTeam)) throw new RuntimeException("팀 가입 후 이용해주세요.");
+				
+		Regist regist = Regist.builder()
+				.sendTeam(Team.builder().teamId(curTeam.getTeamId()).build())
+				.receiveTeam(Team.builder().teamId(otherTeamId).build())
+				.regDate(LocalDateTime.now())
+				.status(2)
+				.build();
 		
-		//TODO : Add error checking mechanism (cannot request to more than one team)
-		
-		return registRepository.save(registReq.toRegist());
+		return registRepository.save(regist);
 	}
 	 
 	
-	//2. 신청 현황 조회 : 나의 팀이 받은 신청과 나의 팀이 보낸 신청을 조회한다.
-	public List<RegistRes> viewReceivedRegistRecords() {
+	//2. 신청 현황 조회 : 나의 팀이 받은 신청을 조회한다.
+	public List<RegistRes> viewReceivedRegistRecords(Integer memberId) {
 		//1) session 에서 나의 팀 아이디를 꺼낸다.
-			Integer curTeamId = (Integer)session.getAttribute("curTeamId");
+		Team curTeam = memberRepository.findByMemberId(memberId)
+				.orElseThrow(() -> new RuntimeException("다시 로그인 해주세요.")).getTeam();
+		if (Objects.isNull(curTeam)) throw new RuntimeException("팀 가입 후 이용해주세요.");
+		
 		//2) DB를 조회해서 나의 팀이 받은 신청을 모두 꺼낸다.
-			List<Regist> registEntities = registRepository.getReceivedRegists(curTeamId);
+		List<Regist> registEntities = registRepository.getReceivedRegists(curTeam.getTeamId());
 		
 		//3) 리스트의 형태로 리턴한다.
-			 return registEntities.stream()
-		                .map(RegistRes::new)
-		                .collect(Collectors.toList());
+		 return registEntities.stream()
+	                .map(RegistRes::new)
+	                .collect(Collectors.toList());
 	}
 	
 	//2.5. 신청 현황 조회 : 나의 팀이 보낸 신청을 조회한다.
-	public List<RegistRes> viewSentRegistRecords(RegistReq registReq) {
+	public List<RegistRes> viewSentRegistRecords(Integer memberId) {
 		//1) session 에서 나의 팀 아이디를 꺼낸다.
-		Integer curTeamId = (Integer)session.getAttribute("curTeamId");
+		Team curTeam = memberRepository.findByMemberId(memberId)
+				.orElseThrow(() -> new RuntimeException("다시 로그인 해주세요.")).getTeam();
+		if (Objects.isNull(curTeam)) throw new RuntimeException("팀 가입 후 이용해주세요.");
+				
 		//2) DB를 조회해서 나의 팀이 받은 신청을 모두 꺼낸다.
-		List<Regist> registEntities = registRepository.getSentRegists(curTeamId);
+		List<Regist> registEntities = registRepository.getSentRegists(curTeam.getTeamId());
 	
 		//3) 리스트의 형태로 리턴한다.
 		 return registEntities.stream()
@@ -62,19 +73,26 @@ public class RegistService {
 	
 	//3. 내가 받은 신청 중 하나를 수락한다.
 	@Transactional
-	public void acceptRegist(RegistReq registReq, Integer curTeamId) throws RuntimeException{
-		Regist regist = registRepository.findById(registReq.getSendTeamId())
+	public void acceptRegist(Integer registId, Integer memberId) throws RuntimeException{
+		Team curTeam = memberRepository.findByMemberId(memberId)
+				.orElseThrow(() -> new RuntimeException("다시 로그인 해주세요.")).getTeam();
+		if (Objects.isNull(curTeam)) throw new RuntimeException("팀 가입 후 이용해주세요.");
+		
+		Regist regist = registRepository.findById(registId)
 				.orElseThrow(()->new RuntimeException("찾을 수 없는 요청입니다."));
+		
+		if (!Objects.equals(regist.getReceiveTeam().getTeamId(), curTeam.getTeamId()))
+			throw new RuntimeException("자신이 소속된 팀의 게스팅만 수락할 수 있습니다..");
 		
 		//수락으로 바꾸기
 		regist.setStatus(0);
 		
 		//내가 받은 요청 모두 불러오기
-		registRepository.getReceivedRegists(curTeamId)
+		registRepository.getReceivedRegists(curTeam.getTeamId())
 				.stream().filter(r->!regist.getRegistId().equals(r.getRegistId())).forEach(r->r.setStatus(1));
 				
 		//내가 보낸 요청 모두 삭제하기
-		List<Regist> sentRegistList =  registRepository.getSentRegists(curTeamId);
+		List<Regist> sentRegistList =  registRepository.getSentRegists(curTeam.getTeamId());
 		registRepository.deleteAll(sentRegistList);
 		
 	}
@@ -82,18 +100,31 @@ public class RegistService {
 	 
 	 //4. 내가 받은 신청 중 하나를 거절한다.
 	@Transactional
-	public void declineRegist(RegistReq registReq, Integer curTeamId) throws RuntimeException{
-		Regist receivedRegist = registRepository.findById(registReq.getSendTeamId())
+	public void declineRegist(Integer registId, Integer memberId) throws RuntimeException{
+		Team curTeam = memberRepository.findByMemberId(memberId)
+				.orElseThrow(() -> new RuntimeException("다시 로그인 해주세요.")).getTeam();
+		if (Objects.isNull(curTeam)) throw new RuntimeException("팀 가입 후 이용해주세요.");
+		
+		Regist regist = registRepository.findById(registId)
 				.orElseThrow(()->new RuntimeException("찾을 수 없는 요청입니다."));
+		
+		if (!Objects.equals(regist.getReceiveTeam().getTeamId(), curTeam.getTeamId()))
+			throw new RuntimeException("자신이 소속된 팀의 게스팅만 거절할 수 있습니다..");
+		
 		//거절로 바꾸기
-		receivedRegist.setStatus(1);
+		regist.setStatus(1);
 	}
 	 
 	 
 	 //5. 성사된 매칭을 조회한다.
-	public Regist getMyMatching (Integer curTeamId) throws RuntimeException{
+	public RegistRes getMyMatching (Integer memberId) throws RuntimeException{
+		Team curTeam = memberRepository.findByMemberId(memberId)
+				.orElseThrow(() -> new RuntimeException("다시 로그인 해주세요.")).getTeam();
+		if (Objects.isNull(curTeam)) throw new RuntimeException("팀 가입 후 이용해주세요.");
+		
 		//내 팀 아이디를 이용해 성사된 매칭을 조회한다.
-		Regist regist = registRepository.getMatchedTeam(curTeamId).orElseThrow();
-		return regist;
+		Regist regist = registRepository.getMatchedTeam(curTeam.getTeamId())
+				.orElseThrow(() -> new RuntimeException("성사된 매칭이 존재하지 않습니다."));
+		return new RegistRes(regist);
 	}
 }
